+++ src/main/kotlin/domain/Dealer.kt
@@ -1,5 +1,7 @@
 package domain
 
+import domain.card.Cards
+
 class Dealer(cards: Cards) : Participant(Name(DEALER_NAME), cards) {
     override fun isPossibleDrawCard(): Boolean = resultSum() <= DEALER_ADD_CARD_CONDITION
