+++ src/main/kotlin/domain/Participant.kt
@@ -1,5 +1,8 @@
 package domain
 
+import domain.card.Card
+import domain.card.Cards
+
 abstract class Participant(val name: Name, val cards: Cards) {
     abstract fun isPossibleDrawCard(): Boolean
     fun addCard(card: Card) = cards.add(card)
