+++ src/main/kotlin/domain/phase/InitPhase.kt
@@ -1,8 +1,8 @@
 package domain.phase
 
-import domain.CardDeck
 import domain.Participants
 import domain.Players
+import domain.card.CardDeck
 
 class InitPhase(val printGameInit: (Players) -> Unit, val printInitCards: (Participants) -> Unit) : Phase {
     override fun runPhase(participants: Participants, deck: CardDeck) {
