+++ src/main/kotlin/domain/phase/Phase.kt
@@ -1,7 +1,7 @@
 package domain.phase
 
-import domain.CardDeck
 import domain.Participants
+import domain.card.CardDeck
 
 interface Phase {
     fun runPhase(participants: Participants, deck: CardDeck)
